requirements.txt
Flask==0.12
Flask-MySQLdb==0.2.0
Flask-SQLAlchem==2.2
itchat==1.3.4
python-redmine==2.0.1
redis==2.10.5
pytest==2.9.2
